# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface StationOverview {
  dataReady: boolean;
  stationsIndexed: number;
  citiesIndexed: number;
  algorithm: string;
  dataset: string;
}

export interface CitySuggestion {
  name: string;
  /** @nullable */
  stateProvince: string | null;
  latitude: number;
  longitude: number;
  stationCount: number;
}

export interface StationResult {
  stationId: string;
  name: string;
  city: string;
  /** @nullable */
  stateProvince: string | null;
  latitude: number;
  longitude: number;
  distanceKm: number;
  /** @nullable */
  ports: number | null;
  /** @nullable */
  powerKw: number | null;
  /** @nullable */
  powerClass: string | null;
  /** @nullable */
  fastDc: boolean | null;
}

export interface NearestStationsResponse {
  stations: StationResult[];
  resultCount: number;
  stationsIndexed: number;
  nodesVisited: number;
  branchesPruned: number;
  searchTimeMs: number;
}

export type GetCitiesParams = {
/**
 * @maxLength 100
 */
q?: string;
};

export type GetNearestStationsParams = {
/**
 * @minimum -90
 * @maximum 90
 */
latitude: number;
/**
 * @minimum -180
 * @maximum 180
 */
longitude: number;
/**
 * @minimum 1
 * @maximum 10
 */
k: number;
};