# lib/api-zod/src/generated/types/citySuggestion.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CitySuggestion {
  name: string;
  /** @nullable */
  stateProvince: string | null;
  latitude: number;
  longitude: number;
  stationCount: number;
}