# lib/api-zod/src/generated/types/nearestStationsResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StationResult } from './stationResult';

export interface NearestStationsResponse {
  stations: StationResult[];
  resultCount: number;
  stationsIndexed: number;
  nodesVisited: number;
  branchesPruned: number;
  searchTimeMs: number;
}