# lib/api-zod/src/generated/types/stationOverview.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StationOverview {
  dataReady: boolean;
  stationsIndexed: number;
  citiesIndexed: number;
  algorithm: string;
  dataset: string;
}