# lib/api-zod/src/generated/types/stationResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StationResult {
  stationId: string;
  name: string;
  city: string;
  /** @nullable */
  stateProvince: string | null;
  latitude: number;
  longitude: number;
  distanceKm: number;
  /** @nullable */
  ports: number | null;
  /** @nullable */
  powerKw: number | null;
  /** @nullable */
  powerClass: string | null;
  /** @nullable */
  fastDc: boolean | null;
}